+++ tests/commands/test_log_command.py
@@ -2,7 +2,7 @@ from src.commands.log_exception_command import LogExceptionCommand
 from src.commands.move import MoveCommand
 from src.exceptions.move import UnchangeablePositionError
 from src.models.vector import Vector
-from tests.test_move import make_movable_object
+from tests.factories import make_movable_object
 
 
 def test_log_exception_command(caplog):
