+++ tests/commands/test_move.py
@@ -1,17 +1,9 @@
 import pytest
 
-from src.adapters.movable_adapter import MovableObjectAdapter
 from src.commands.move import MoveCommand
 from src.exceptions.move import UndefinedPositionError, UndefinedVelocityError, UnchangeablePositionError
 from src.models.vector import Vector
-from tests.mock_object import MockUObject
-
-
-def make_movable_object(position: Vector, velocity: Vector) -> MovableObjectAdapter:
-    data = {"position": position, "velocity": velocity}
-    mock_object = MockUObject(data)
-    movable_object = MovableObjectAdapter(mock_object)
-    return movable_object
+from tests.factories import make_movable_object
 
 
 def test_move_with_valid_params():
